+++ src/auth/auth.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
-import { FirebaseModule } from 'src/firebase/firebase.module';
-import { FirebaseAuthStrategy } from 'src/firebase/firebase-auth.strategy';
+import { FirebaseModule } from '../firebase/firebase.module';
+import { FirebaseAuthStrategy } from '../firebase/firebase-auth.strategy';
 
 @Module({
   imports: [FirebaseModule],
