+++ src/auth/auth.service.ts
@@ -1,6 +1,6 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { FIREBASE } from 'src/firebase/firebase-factory';
-import { FirebaseServices } from 'src/firebase/firebase.types';
+import { FIREBASE } from '../firebase/firebase-factory';
+import { FirebaseServices } from '../firebase/firebase.types';
 import { UserAuthDto } from './dto/user-auth.dto';
 
 @Injectable()
