+++ src/partner-access/partner-access.interface.ts
