+++ src/partner-admin/partner-admin.interface.ts
