+++ src/partner/partner.interface.ts
